+++ actix-web/src/http/header/content_range.rs
@@ -67,7 +67,6 @@ crate::http::header::common_header! {
         crate::http::header::common_header_test!(test_bytes_many_dashes,
             [b"bytes 1-2-3/500"],
             None::<ContentRange>);
-
     }
 }