+++ actix-web/CHANGES.md
@@ -8,6 +8,7 @@
 - Add `Resource::{get, post, etc...}` methods for more concisely adding routes that don't need additional guards.
 - Add `web::Payload::to_bytes[_limited]()` helper methods.
 - Add missing constructors on `HttpResponse` for several status codes.
+- Add `http::header::ContentLength` typed header.
 
 ### Changed